==> scATAC-seq/2-QC/5-batch_effect_correction/2-data_integration_multiome/03-normalization_and_integration_with_multiome.Rmd <==
---
title: "Normalization and Batch Effect Correction Integration Data"
author: "Paula Soler-Vila"
output: 
  BiocStyle::html_document:
    toc: true
    toc_float: true
    number_sections: true
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, out.width = "100%", fig.align='center', 
                      message=FALSE, warning = FALSE)
options(width = 1200)
```


# Introduction

In previous notebooks and scripts, we integrated our scATAC dataset, and now we are going to merge it with the ATAC fraction of the multiome dataset. Here, we will correct for batch effects with [Harmony](https://www.nature.com/articles/s41592-019-0619-0), and visualize the intermixing of different potential confounders pre- and post-integration.

In addition, we will save the dimensionality reduction (PCA) matrices before and after integration to further quantify the effect of the aforementioned confounders in high dimensional space. 


# Pre-processing

## Load packages

```{r}
library(Seurat)
library(SeuratWrappers)
library(Signac)
library(harmony)
library(tidyverse)

set.seed(222)
```


## Parameters

```{r}
# Paths
path_to_obj <- here::here("scATAC-seq/results/R_objects/6.tonsil_atac_merged_with_multiome_missing_integration.rds")
path_to_save_obj <- here::here("scATAC-seq/results/R_objects/7.tonsil_atac_integrated_with_multiome.rds")
path_tmp_dir <- here::here("scATAC-seq/2-QC/5-batch_effect_correction/2-data_integration_multiome/tmp/")

path_to_save_dimred_uncorrect <- str_c(path_tmp_dir, "batch_uncorrected_lsi.rds", sep = "")
path_to_save_dimred_correct <- str_c(path_tmp_dir, "batch_corrected_lsi.rds", sep = "")
path_to_save_confounders_df <- str_c(path_tmp_dir, "confounders_df.rds", sep = "") 
```


## Load data

```{r}
tonsil <- readRDS(path_to_obj)
tonsil
```

# Visualize UMAP without batch effect correction

```{r fig.wide=TRUE}
# Process Seurat object
tonsil <- tonsil %>%
  RunTFIDF() %>% 
  FindTopFeatures(min.cutoff = "q0") %>%
  RunSVD() %>%
  RunUMAP(reduction = "lsi", dims = 2:40)

DepthCor(tonsil)

# Visualize UMAP
confounders <- c("library_name", "sex", "age_group", "hospital", "assay")
umaps_before_integration <- purrr::map(confounders, function(x) {
  p <- DimPlot(tonsil, group.by = x, pt.size = 0.1)
  p
})
names(umaps_before_integration) <- confounders
print("UMAP colored by GEM:")
umaps_before_integration$library_name + NoLegend()
print("UMAP colored by sex, age group, cell hashing status, sampling center and assay:")
umaps_before_integration[2:length(umaps_before_integration)]
```


# Run and visualize Harmony's integration

```{r fig.wide=TRUE}
tonsil <- RunHarmony(
  object = tonsil,
  group.by.vars = "gem_id", 
  reduction = "lsi",
  dims = 2:40,
  assay.use = "peaks",
  project.dim = FALSE
)

tonsil <- RunUMAP(tonsil, dims = 2:40, reduction = "harmony")

umaps_after_integration <- purrr::map(confounders, function(x) {
  p <- DimPlot(tonsil, group.by = x, pt.size = 0.1)
  p
})
names(umaps_after_integration) <- confounders
print("UMAP colored by GEM:")
umaps_after_integration$library_name + NoLegend()
print("UMAP colored by sex, age group, cell hashing status, sampling center and assay:")
umaps_after_integration[2:length(umaps_before_integration)]
```

# Save

```{r}
# If it doesn't exist create temporal directory
#dir.create(path_tmp_dir, showWarnings = FALSE) 


# Save integrated Seurat object
saveRDS(tonsil, path_to_save_obj)


# Save PCA matrices to compute the Local Inverse Simpson Index (LISI)
confounders_df <- tonsil@meta.data[, confounders]
saveRDS(confounders_df, path_to_save_confounders_df)
saveRDS(
  tonsil@reductions$lsi@cell.embeddings[, 2:40],
  path_to_save_dimred_uncorrect
)
saveRDS(
  tonsil@reductions$harmony@cell.embeddings[, 2:40],
  path_to_save_dimred_correct
)
```


# Session Information

```{r}
sessionInfo()
```
